Refs #8910 - Change Fortran to support EMU.
Note that this only involved a straight swap of any occurances of
"64" in the code with "96".  This was never meant to be a future-
proof method of supporting more than the previously-supported 64
detectors.

This change will require testing from the scientists to verify that
things are still working as expected.