Minor amendments to library target
To be able to build we need gfortran. On versions > 10 we need to turn off warnings of COMPLEX TO REAL conversions, as it stops us being able to build (the warnings are now errors). This commit makes sure we can find gfortran and supplies the compiler with the correct arguments.